API Reference

Public API

OptimizingIR — Module

An Intermediate Representation (IR) on steroids.

Internals

OptimizingIR.AbstractValue — Type

AbstractValue{M<:Mutability}

A value can be marked as either Mutable or Immutable.

An immutable value can be assigned only once. A mutable value can be assigned more than once.

OptimizingIR.PureInstruction — Type

A PureInstruction is a call to an operation that always returns the same value if the same arguments are passed to the instruction. It is suitable for memoization, in the sense that it can be optimized in the Value-Number algorithm inside a Basic Block.

OptimizingIR.ImpureInstruction — Type

An ImpureInstruction is a call to an operation that not always returns the same value if the same arguments are passed to the instruction. It is not suitable for memoization, and the Value-Number optimization must be disabled for this call.

Marking as mutable avoids Value-Numbering for this call.

OptimizingIR.compile — Function

compile(::Type{T}, program::Program) :: Function where {T<:AbstractMachine}

OptimizingIR.SSAValue — Type

A pointer to an instruction that computes a value.

OptimizingIR.try_on_add_instruction_passes — Function

try_on_add_instruction_passes(program, instruction) :: Union{Nothing, Address}

Tries to apply all optimization passes available while running addinstruction!:

function addinstruction!(b::BasicBlock, instruction::LinearInstruction) :: Address

    result = try_on_add_instruction_passes(b, instruction)
    if result != nothing
        return result
    end

    # (...)
end

OptimizingIR.LookupTable — Type

Generic struct for a lookup table that stores an ordered list of distinct elements.

element is stored in entries vector at index i.

index[element] retrieves the index i.

Use addentry! to add items to the table. It the table already has the item, addentry! will return the existing item's index.

OptimizingIR.Const — Type

Constant value to be encoded directly into the IR. The address is the value itself.

OptimizingIR.Variable — Type

A variable that can be binded to a value. See OptimizingIR.bind!.

OptimizingIR.OptimizationRule — Type

Sets which optimizations are allowed to an Op.