API Reference

Public API

OptimizingIR.Op — Type

Op(f::Function;
        pure::Bool=false,
        commutative::Bool=false,
        hasleftidentity::Bool=false,
        hasrightidentity::Bool=false,
        identity_element::T=NULL_IDENTITY_ELEMENT)

Defines a basic instruction with optimization annotations.

Arguments

f is a Julia function to be executed by the Op.

pure: marks the function as pure (true) or impure (false) .

commutative: marks the Op as commutative.

hasleftidentity: marks the Op as having an identity when operating from the left, which means that f(I, v) = v, where I is the identity_element.

hasrightidentity: marks the Op as having an identity when operating from the right, which means that f(v, I) = v, where I is the identity_element.

Purity

A function is considered pure if its return value is the same for the same arguments, and has no side-effects.

Operations marked as pure are suitable for Value numbering optimization.

When marked as impure, all optimization passes are disabled.

Examples

const op_sum = OIR.Op(+, pure=true, commutative=true, hasleftidentity=true, hasrightidentity=true, identity_element=0)
const op_sub = OIR.Op(-, pure=true, hasrightidentity=true, identity_element=0)
const op_mul = OIR.Op(*, pure=true, commutative=true, hasleftidentity=true, hasrightidentity=true, identity_element=1)
const op_div = OIR.Op(/, pure=true, hasrightidentity=true, identity_element=1)
const op_pow = OIR.Op(^, pure=true, hasrightidentity=true, identity_element=1)

foreign_fun(a, b, c) = a^3 + b^2 + c
const op_foreign_fun = OIR.Op(foreign_fun, pure=true)

const op_zeros = OIR.Op(zeros)

OptimizingIR.Variable — Type

Variable

Creates a variable identified by a symbol that can be either mutable or immutable.

Examples

m = OptimizingIR.MutableVariable(:varmut) # a mutable variable
im = OptimizingIR.ImmutableVariable(:varimut) # an immutable variable

OptimizingIR.call — Function

call(op, args...) :: LinearInstruction

Creates an instruction as a call to operation op with arguments args.

OptimizingIR.addinstruction! — Function

addinstruction!(b::BasicBlock, instruction::LinearInstruction) :: ImmutableValue

Pushed an instruction to a basic block. Returns the value that represents the result after the execution of the instruction.

OptimizingIR.addinput! — Function

addinput!(b::BasicBlock, iv::ImmutableVariable)

Registers iv as an input variable of the function.

OptimizingIR.addoutput! — Function

addoutput!(b::BasicBlock, iv::Variable)

Registers iv as an output variable of the function.

OptimizingIR.callgetindex — Function

callgetindex(array, index...) :: LinearInstruction

Creates an instruction that results in the value of array at index.

OptimizingIR.callsetindex — Function

callsetindex(array, value, index...) :: LinearInstruction

Creates an instruction that sets value in array at index.

OptimizingIR.assign! — Function

assign!(bb::BasicBlock, lhs::Variable, rhs::AbstractValue)

Assigns the value rhs to variable lhs.

OptimizingIR.constant — Function

constant(val) :: Const

Creates a constant value.

OptimizingIR.compile — Function

compile(::Type{T}, program::Program) :: Function where {T<:AbstractMachine}

Compiles the IR to a Julia function.

const OIR = OptimizingIR
bb = OIR.BasicBlock()
# (...) add instructions to basic block
finterpreter = OIR.compile(OIR.BasicBlockInterpreter, bb)
fnative = OIR.compile(OIR.Native, bb)

Internals

OptimizingIR.AbstractValue — Type

AbstractValue{M<:Mutability}

A value can be marked as either Mutable or Immutable.

An immutable value can be assigned only once. A mutable value can be assigned more than once.

OptimizingIR.PureInstruction — Type

A PureInstruction is a call to an operation that always returns the same value if the same arguments are passed to the instruction. It is suitable for memoization, in the sense that it can be optimized in the Value-Number algorithm inside a Basic Block.

OptimizingIR.ImpureInstruction — Type

An ImpureInstruction is a call to an operation that not always returns the same value if the same arguments are passed to the instruction. It is not suitable for memoization, and the Value-Number optimization must be disabled for this call.

Marking as mutable avoids Value-Numbering for this call.

OptimizingIR.SSAValue — Type

A pointer to an instruction that computes a value.

OptimizingIR.try_on_add_instruction_passes — Function

try_on_add_instruction_passes(program, instruction) :: Union{Nothing, ImmutableValue}

Tries to apply all optimization passes available while running addinstruction!:

function addinstruction!(b::BasicBlock, instruction::LinearInstruction) :: ImmutableValue

    result = try_on_add_instruction_passes(b, instruction)
    if result != nothing
        return result
    end

    # (...)
end

OptimizingIR.LookupTable — Type

Generic struct for a lookup table that stores an ordered list of distinct elements.

element is stored in entries vector at index i.

index[element] retrieves the index i.

Use addentry! to add items to the table. It the table already has the item, addentry! will return the existing item's index.

OptimizingIR.Const — Type

Constant value to be encoded directly into the IR. The address is the value itself.

OptimizingIR.OptimizationRule — Type

Sets which optimizations are allowed to an Op.